sse handler apart
