+++ example/sse/sse-server.js
@@ -1,0 +1,10 @@
+const sseRequestHandler = (req, res) => {
+  res.set("Content-Type", "text/event-stream");
+
+  const data = { hello: "world" };
+
+  res.write(`data: ${JSON.stringify(data)}`);
+  res.write("\n\n");
+};
+
+module.exports = sseRequestHandler;
